requirements.txt
pandas>=2.0
numpy>=1.24
kafka-python>=2.0
datasets>=2.14
